+++ PopdeemSDK/UIKit/Common/PDUIInboxButton.m
@@ -34,7 +34,7 @@
 		UILabel *lbl_card_count = [[UILabel alloc]initWithFrame:topRight];
 		lbl_card_count.textColor = [UIColor whiteColor];
 		lbl_card_count.textAlignment = NSTextAlignmentCenter;
-		lbl_card_count.text = [NSString stringWithFormat:@"%ld",unread];
+    lbl_card_count.text = [NSString stringWithFormat:@"%ld",(unsigned long)unread];
 		lbl_card_count.layer.borderWidth = 1;
 		lbl_card_count.layer.cornerRadius = 8;
 		lbl_card_count.layer.masksToBounds = YES;
